pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zoom-mcp"
version = "0.1.0"
description = "Zoom MCP Server for room status monitoring"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]

requires-python = ">=3.10"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
zoom-mcp = "src.server:main"